Login page added in development

+++ Python UI/UI.py
@@ -849,44 +849,52 @@ def GUI_Server_Comms():
 
 app = Flask(__name__) #main web application
 
+logged_in = True
+
 #main page of the website
 @app.route("/", methods=['GET', 'POST']) # methods of interating and route address
 def Main_page():
-    # Reads all the current commands in the buffer and addes N/A if blank
-    Next =[] # list of next 4 commands
-    for i in range(0, 4):
+    if logged_in == False:
+        template = {
+            'address': '/',
+        }
+        return render_template('login.html', **template) #Renders webpage 
+    else:
+        # Reads all the current commands in the buffer and addes N/A if blank
+        Next =[] # list of next 4 commands
+        for i in range(0, 4):
+            try: 
+                Next.append(N_CMD.get(block = False)) # reads queue and adds commands to the list.
+            except: 
+                Next.append("N/A") # N/A if it is not available
         try: 
-            Next.append(N_CMD.get(block = False)) # reads queue and adds commands to the list.
+            Current = C_CMD.get(block = False) #Reads current command 
         except: 
-            Next.append("N/A") # N/A if it is not available
-    try: 
-        Current = C_CMD.get(block = False) #Reads current command 
-    except: 
-        Current = "N/A"
+            Current = "N/A"
 
-    #To edit the webpage HTML file with the python data
-    template = {
-        'title': 'Arcadius Monitoring Page',
-        'Current' : Current,
-        'N1' : Next[0], #the next 4 commands
-        'N2' : Next[1],
-        'N3' : Next[2],
-        'N4' : Next[3]
-    }
-    
-    # Gets the request form the webpage about button presses
-    if request.method == 'POST':
-            if request.form.get('Kill') == 'Kill': # if the form item called Kill is clicked it reads the value
-                Kill_Conn.send("kill") # Sends kill command on kill pipe
-            elif request.form.get('Screenshot') == 'Screenshot':
-                frame = web_frame.get()
-                file_name = str(datetime.datetime.now()).replace('.','_').replace('-','_').replace(':','_') + ".jpg"
-                cv2.imwrite(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screenshots\\', file_name), frame)
-            else:
-                pass
-    elif request.method == 'GET':
-        pass
-    return render_template('Main.html', **template) #Renders webpage 
+        #To edit the webpage HTML file with the python data
+        template = {
+            'title': 'Arcadius Monitoring Page',
+            'Current' : Current,
+            'N1' : Next[0], #the next 4 commands
+            'N2' : Next[1],
+            'N3' : Next[2],
+            'N4' : Next[3]
+        }
+        
+        # Gets the request form the webpage about button presses
+        if request.method == 'POST':
+                if request.form.get('Kill') == 'Kill': # if the form item called Kill is clicked it reads the value
+                    Kill_Conn.send("kill") # Sends kill command on kill pipe
+                elif request.form.get('Screenshot') == 'Screenshot':
+                    frame = web_frame.get()
+                    file_name = str(datetime.datetime.now()).replace('.','_').replace('-','_').replace(':','_') + ".jpg"
+                    cv2.imwrite(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screenshots\\', file_name), frame)
+                else:
+                    pass
+        elif request.method == 'GET':
+            pass
+        return render_template('Main.html', **template) #Renders webpage 
 
 # generates videofeed from frame queue
 def gen_frames(): 
@@ -906,32 +914,44 @@ def gen_frames():
 #Renders a webpage fro pure video streaming which is linked to on main page
 @app.route('/video_feed')
 def video_feed():
-    return Response(gen_frames(), mimetype='multipart/x-mixed-replace; boundary=frame')
+    if logged_in == False:
+        template = {
+            'address': '/video_feed',
+        }
+        return render_template('login.html', **template) #Renders webpage 
+    else:
+        return Response(gen_frames(), mimetype='multipart/x-mixed-replace; boundary=frame')
 
 #New control page
 @app.route('/control', methods=['GET', 'POST'])
 def control_page():
-    if request.method == 'POST':
-            if request.form.get('Kill') == 'Kill': # if the form item called Kill is clicked it reads the value
-                Kill_Conn.send("kill") # Sends kill command on kill pipe
-                return render_template('control.html')
-            elif request.form.get('Screenshot') == 'Screenshot':
-                frame = web_frame.get() 
-                file_name = str(datetime.datetime.now()).replace('.','_').replace('-','_').replace(':','_')  + ".jpg"
-                cv2.imwrite(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screenshots\\', file_name), frame)
-            elif  request.form.get('Pump') == 'Pump':
-                print("Pump")
-                CMD_Conn.send("PUMP")
-                return render_template('control.html')
-            elif  request.form.get('Shutter') == 'Shutter':
-                print("Shutter")
-                CMD_Conn.send("Shutter")
-                return render_template('control.html')
-            else:
-                pass
-    elif request.method == 'GET':
-        pass
-    return render_template('control.html')
+    if logged_in == False:
+        template = {
+            'address': '/control',
+        }
+        return render_template('login.html', **template) #Renders webpage 
+    else:
+        if request.method == 'POST':
+                if request.form.get('Kill') == 'Kill': # if the form item called Kill is clicked it reads the value
+                    Kill_Conn.send("kill") # Sends kill command on kill pipe
+                    return render_template('control.html')
+                elif request.form.get('Screenshot') == 'Screenshot':
+                    frame = web_frame.get() 
+                    file_name = str(datetime.datetime.now()).replace('.','_').replace('-','_').replace(':','_')  + ".jpg"
+                    cv2.imwrite(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screenshots\\', file_name), frame)
+                elif  request.form.get('Pump') == 'Pump':
+                    print("Pump")
+                    CMD_Conn.send("PUMP")
+                    return render_template('control.html')
+                elif  request.form.get('Shutter') == 'Shutter':
+                    print("Shutter")
+                    CMD_Conn.send("Shutter")
+                    return render_template('control.html')
+                else:
+                    pass
+        elif request.method == 'GET':
+            pass
+        return render_template('control.html')
 
 #---------- GUI Thread -------------#
 def GUI():
